requirements.txt
mcp
aiohttp
sqlalchemy[asyncio]
asyncpg  # For PostgreSQL
aiomysql  # For MySQL
stripe
pymongo
twilio
sendgrid
hubspot-api-client
simple-salesforce
notion-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
O365
tweepy
python-dotenv